put services in module and add tests
