+++ app/models/clockwork.rb
@@ -48,11 +48,11 @@ module Clockwork
   if Rails.env == 'production'
     every(1.day, 'generate_quarterly_reports', at: '01:00', if: -> (t) { t.day == 20 && (t.month == 1 || t.month == 4 || t.month == 7 || t.month == 10) }) do
       Antenne.find_each do |antenne|
-        QuarterlyReportService.new(antenne).delay(queue: :low_priority).call
+        QuarterlyReports::GenerateReports.new(antenne).delay(queue: :low_priority).call
       end
     end
     every(1.day, 'send_quarterly_reports_emails', at: '08:00', if: -> (t) { t.day == 21 && (t.month == 1 || t.month == 4 || t.month == 7 || t.month == 10) }) do
-      SendManagersEmailsService.new.delay(queue: :low_priority).call
+      QuarterlyReports::NotifyManagers.new.delay(queue: :low_priority).call
     end
     every(1.day, 'reminders_registers', :at => ['01:00', '13:00']) { RemindersService.delay(queue: :low_priority).create_reminders_registers }
   end
