--- app/services/send_managers_emails_service.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-module SendManagersEmailsService
-  def initialize(managers = User.managers)
-    @managers = managers
-  end
-
-  def call
-    @managers.each do |user|
-      next unless user.managed_antennes.map { |antenne| antenne.quarterly_reports.find_by(start_date: 3.months.ago.beginning_of_month) }.any?
-      UserMailer.quarterly_report(user).deliver_later
-    end
-  end
-end
