+++ spec/services/quarterly_reports/notify_managers_spec.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+describe QuarterlyReports::NotifyManagers do
+  describe 'call' do
+    let(:antenne) { create :antenne }
+    let!(:quarterly_report) { create :quarterly_report, antenne: antenne, start_date: 3.months.ago.beginning_of_month }
+    let!(:manager) { create :user, :manager, managed_antennes: [antenne] }
+    let!(:normal_user) { create :user }
+
+    subject { described_class.new.call }
+
+    it 'send a mail to managers' do
+      expect { subject }.to change { ActionMailer::Base.deliveries.count }.by(1)
+    end
+  end
+end
